lanacc_project/lanacc/python/requirements.txt
openpyxl==3.1.2
mysql-connector-python==8.3.0
pandas==2.2.1
python-dotenv==1.0.1
rich==13.7.1
click==8.1.7
